Label to recourse & transfer MainActivity.kt
