+++ app/src/main/java/ru/konovalovk/translateplayer/ui/MainActivity.kt
@@ -1,0 +1,18 @@
+package ru.konovalovk.translateplayer.ui
+
+import android.os.Bundle
+import android.view.WindowManager
+import androidx.appcompat.app.AppCompatActivity
+import ru.konovalovk.translateplayer.R
+
+
+class MainActivity : AppCompatActivity() {
+    val TAG = this::class.java.simpleName
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_main)
+        window.addFlags(WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON)
+
+    }
+}
